requirements.txt
tensorflow
pandas 
numpy
scikit-image
scikit-learn
dvc
mlflow
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
scipy
Flask
Flask-Cors
gdown

